hwm-events/src/requirements.txt
google-cloud-gdchardwaremanagement
google-cloud-firestore
google-cloud-pubsub
google-auth
functions-framework
